What: switch config codec from json to yaml
Why:

  * more clear

How:

  * nop
