+++ cmd/gogo/templates/application_config.go
@@ -9,26 +9,27 @@ import (
 	"{{.Namespace}}/{{.Application}}/app/models"
 )
 
+// Shared Config
 var (
 	Config *AppConfig
 )
 
 // AppConfig defines specs for config
 type AppConfig struct {
-	Model        *models.Config      ` + "`" + `json:"model"` + "`" + `
-	Domain       string              ` + "`" + `json:"domain"` + "`" + `
-	GettingStart *GettingStartConfig ` + "`" + `json:"getting_start"` + "`" + `
-	Debug        bool                ` + "`" + `json:"debug"` + "`" + `
+	Model        *models.Config      ` + "`" + `yaml:"model"` + "`" + `
+	Domain       string              ` + "`" + `yaml:"domain"` + "`" + `
+	GettingStart *GettingStartConfig ` + "`" + `yaml:"getting_start"` + "`" + `
+	Debug        bool                ` + "`" + `yaml:"debug"` + "`" + `
 }
 
 // NewAppConfig creates Config from gogo.Configer
 func NewAppConfig(config gogo.Configer) error {
-	return config.UnmarshalJSON(&Config)
+	return config.UnmarshalYAML(&Config)
 }
 
-// Sample config for illustration
+// GettingStartConfig is a sample config for illustration
 type GettingStartConfig struct {
-	Greeting string ` + "`" + `json:"greeting"` + "`" + `
+	Greeting string ` + "`" + `yaml:"greeting"` + "`" + `
 }
 `
 	applicationConfigTestTemplate = `package controllers
@@ -46,74 +47,52 @@ func Test_AppConfig(t *testing.T) {
 	it.NotNil(Config.GettingStart)
 }
 `
-	applicationConfigJSONTemplate = `{
-	"mode": "test",
-	"name": "{{.Application}}",
-	"sections": {
-		"development": {
-			"server": {
-				"addr": "localhost",
-				"port": 9090,
-				"healthz": true,
-				"ssl": false,
-				"request_timeout": 3,
-				"response_timeout": 10,
-				"request_id": "X-Request-Id"
-			},
-			"logger": {
-				"output": "stderr",
-				"level": "debug",
-				"filter_params": ["password", "password_confirmation"]
-			},
-			"domain": "https://example.com",
-			"getting_start": {
-				"greeting": "Hello, gogo!"
-			},
-			"debug": true
-		},
+	applicationConfigYAMLTemplate = `---
+mode: test
+name: {{.Application}}
 
-		"test": {
-			"server": {
-				"addr": "localhost",
-				"port": 9090,
-				"ssl": false,
-				"request_timeout": 3,
-				"response_timeout": 10,
-				"request_id": "X-Request-Id"
-			},
-			"logger": {
-				"output": "nil",
-				"level": "info",
-				"filter_params": ["password", "password_confirmation"]
-			},
-			"domain": "https://example.com",
-			"getting_start": {
-				"greeting": "Hello, gogo!"
-			},
-			"debug": false
-		},
+default_server: &default_server
+	addr: localhost
+	port: 9090
+	ssl: false
+	request_timeout: 3
+	response_timeout: 10
+	request_id: X-Request-Id
 
-		"production": {
-			"server": {
-				"addr": "localhost",
-				"port": 9090,
-				"healthz": true,
-				"ssl": true,
-				"ssl_cert": "/path/to/ssl/cert",
-				"ssl_key": "/path/to/ssl/key",
-				"throttle": 3000,
-				"demotion": 6000,
-				"request_timeout": 3,
-				"response_timeout": 10,
-				"request_id": "X-Request-Id"
-			},
-			"logger": {
-				"output": "stderr",
-				"level": "warn",
-				"filter_params": ["password", "password_confirmation"]
-			}
-		}
-	}
-}
+default_logger: &default_logger
+	output: nil
+	level: debug
+	filter_params:
+	- password
+	- password_confirmation
+
+sections:
+	development:
+		server:
+			<<: *default_server
+		logger:
+			<<: *default_logger
+		domain: https://example.com
+		getting_start:
+			greeting: Hello, gogo!
+		debug: true
+	test:
+		server:
+			<<: *default_server
+			request_id: ''
+		logger:
+			<<: *default_logger
+		domain: https://example.com
+		getting_start:
+			greeting: Hello, gogo!
+		debug: false
+	production:
+		server:
+			<<: *default_server
+			ssl: true
+			ssl_cert: "/path/to/ssl/cert"
+			ssl_key: "/path/to/ssl/key"
+		logger:
+			<<: *default_logger	
 `
 )
