+++ cmd/gogo/templates/main.go
@@ -25,7 +25,7 @@ var (
 
 func main() {
 	flag.StringVar(&runMode, "runMode", "development", "{{.Application}} -runMode=[development|test|production]")
-	flag.StringVar(&cfgPath, "cfgPath", "", "{{.Application}} -cfgPath=/path/to/[config/application.json]")
+	flag.StringVar(&cfgPath, "cfgPath", "", "{{.Application}} -cfgPath=/path/to/[config/application.yml]")
 	flag.Parse()
 
 	// verify run mode
