+++ internal/render/consts.go
@@ -1,9 +1,11 @@
 package render
 
+// content-types
 const (
-	ContentTypeDefault = "text/plain; charset=utf-8"
-	ContentTypeJSON    = "application/json"
-	ContentTypeJSONP   = "application/javascript"
-	ContentTypeXML     = "text/xml"
-	ContentTypeGRPC    = "application/x-grpc"
+	ContentTypeDefault   = "text/plain; charset=utf-8"
+	ContentTypeJSON      = "application/json"
+	ContentTypeJSONP     = "application/javascript"
+	ContentTypeXML       = "text/xml"
+	ContentTypeProtobuf  = "application/protobuf"
+	ContentTypeProtoJSON = "application/jsonpb"
 )
