+++ internal/render/errors.go
@@ -2,6 +2,7 @@ package render
 
 import "errors"
 
+// errors
 var (
-	ErrHash = errors.New("The hash function does not linked into the binary!")
+	ErrHash = errors.New("The hash function does not linked into the binary")
 )
