+++ internal/render/etag_test.go
@@ -1,8 +1,8 @@
 package render
 
 import (
+	"crypto"
 	"encoding/xml"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -11,63 +11,43 @@ import (
 	"github.com/golib/assert"
 )
 
-func Test_DefaultRender(t *testing.T) {
+func Test_HashRender(t *testing.T) {
 	it := assert.New(t)
 	recorder := httptest.NewRecorder()
 
 	// render with normal string
 	s := "Hello, world!"
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 	it.Equal(ContentTypeDefault, render.ContentType())
 
 	err := render.Render(s)
 	if it.Nil(err) {
 		it.Equal(http.StatusOK, recorder.Code)
-		it.Equal(ContentTypeDefault, recorder.Header().Get("Content-Type"))
+		it.Equal("6cd3556deb0da54bca060b4c39479839", recorder.Header().Get("Etag"))
 		it.Equal(s, recorder.Body.String())
 	}
 }
 
-func Benchmark_DefaultRender(b *testing.B) {
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	recorder := httptest.NewRecorder()
-
-	s := "Hello, world!"
-
-	render := NewDefaultRender(recorder)
-	for i := 0; i < b.N; i++ {
-		recorder.Body.Reset()
-
-		render.Render(s)
-	}
-}
-
-func Test_DefaultRenderWithReader(t *testing.T) {
+func Test_HashRenderWithReader(t *testing.T) {
 	it := assert.New(t)
 	recorder := httptest.NewRecorder()
 
-	s := "Hello, world!"
+	// render with io.Reader
+	reader := strings.NewReader("Hello, world!")
 
-	recorder.Header().Add("Content-Length", fmt.Sprintf("%d", len(s)))
-	recorder.Header().Add("Content-Type", "text/plain")
-
-	// render with normal string
-	reader := strings.NewReader(s)
-
-	render := NewDefaultRender(recorder)
-	it.Equal("text/plain", render.ContentType())
+	render := NewHashRender(recorder, crypto.MD5)
+	it.Equal(ContentTypeDefault, render.ContentType())
 
 	err := render.Render(reader)
 	if it.Nil(err) {
 		it.Equal(http.StatusOK, recorder.Code)
-		it.Equal(s, recorder.Body.String())
+		it.Equal("6cd3556deb0da54bca060b4c39479839", recorder.Header().Get("Etag"))
+		it.Equal("Hello, world!", recorder.Body.String())
 	}
 }
 
-func Benchmark_DefaultRenderWithReader(b *testing.B) {
+func Benchmark_HashRenderWithReader(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 
@@ -75,7 +55,7 @@ func Benchmark_DefaultRenderWithReader(b *testing.B) {
 
 	reader := strings.NewReader("Hello, world!")
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 	for i := 0; i < b.N; i++ {
 		recorder.Body.Reset()
 
@@ -83,7 +63,7 @@ func Benchmark_DefaultRenderWithReader(b *testing.B) {
 	}
 }
 
-func Test_DefaultRenderWithJson(t *testing.T) {
+func Test_HashRenderWithJson(t *testing.T) {
 	it := assert.New(t)
 	recorder := httptest.NewRecorder()
 	recorder.Header().Set("Content-Type", "application/json")
@@ -94,15 +74,16 @@ func Test_DefaultRenderWithJson(t *testing.T) {
 		Age  int
 	}{"gogo", 5}
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 
 	err := render.Render(data)
 	if it.Nil(err) {
+		it.Equal("54843ae1dec66f4fefe6dfa7bcdf1567", recorder.Header().Get("Etag"))
 		it.Equal(`{"Name":"gogo","Age":5}`, strings.TrimSpace(recorder.Body.String()))
 	}
 }
 
-func Benchmark_DefaultRenderWithJson(b *testing.B) {
+func Benchmark_HashRenderWithJson(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 
@@ -114,7 +95,7 @@ func Benchmark_DefaultRenderWithJson(b *testing.B) {
 		Age  int
 	}{"gogo", 5}
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 	for i := 0; i < b.N; i++ {
 		recorder.Body.Reset()
 
@@ -122,8 +103,9 @@ func Benchmark_DefaultRenderWithJson(b *testing.B) {
 	}
 }
 
-func Test_DefaultRenderWithXml(t *testing.T) {
+func Test_HashRenderWithXml(t *testing.T) {
 	it := assert.New(t)
+
 	recorder := httptest.NewRecorder()
 	recorder.Header().Set("Content-Type", "text/xml")
 
@@ -137,15 +119,16 @@ func Test_DefaultRenderWithXml(t *testing.T) {
 		Content: "Hello, world!",
 	}
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 
 	err := render.Render(data)
 	if it.Nil(err) {
+		it.Equal("65693ee59f678f04bc8bedf16f980f5a", recorder.Header().Get("Etag"))
 		it.Equal("<recorder><Result><Success>true</Success><Content>Hello, world!</Content></Result></recorder>", recorder.Body.String())
 	}
 }
 
-func Benchmark_DefaultRenderWithXml(b *testing.B) {
+func Benchmark_HashRenderWithXml(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 
@@ -161,7 +144,7 @@ func Benchmark_DefaultRenderWithXml(b *testing.B) {
 		Content: "Hello, world!",
 	}
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 	for i := 0; i < b.N; i++ {
 		recorder.Body.Reset()
 
@@ -169,7 +152,7 @@ func Benchmark_DefaultRenderWithXml(b *testing.B) {
 	}
 }
 
-func Test_DefaultRenderWithStringify(t *testing.T) {
+func Test_HashRenderWithStringify(t *testing.T) {
 	it := assert.New(t)
 	recorder := httptest.NewRecorder()
 
@@ -179,15 +162,16 @@ func Test_DefaultRenderWithStringify(t *testing.T) {
 		Age  int
 	}{"gogo", 5}
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 
 	err := render.Render(data)
 	if it.Nil(err) {
-		it.Equal(`{gogo 5}`, recorder.Body.String())
+		it.Equal("1b9f54d6753f2e8e4d4a819a44d90ce1", recorder.Header().Get("Etag"))
+		it.Contains(recorder.Body.String(), `{gogo 5}`)
 	}
 }
 
-func Benchmark_DefaultRenderWithStringify(b *testing.B) {
+func Benchmark_HashRenderWithStringify(b *testing.B) {
 	b.ReportAllocs()
 	b.ResetTimer()
 
@@ -198,7 +182,7 @@ func Benchmark_DefaultRenderWithStringify(b *testing.B) {
 		Age  int
 	}{"gogo", 5}
 
-	render := NewDefaultRender(recorder)
+	render := NewHashRender(recorder, crypto.MD5)
 	for i := 0; i < b.N; i++ {
 		recorder.Body.Reset()
