+++ middleware_test.go
@@ -1,5 +1,3 @@
-// +build !race
-
 package gogo
 
 import (
