+++ server_non_race_test.go
@@ -3,172 +3,22 @@
 package gogo
 
 import (
-	"context"
 	"fmt"
 	"io/ioutil"
 	"net"
 	"net/http"
-	"os"
 	"strconv"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
-	"github.com/dolab/httptesting"
 	"github.com/golib/assert"
 )
-
-func Test_ServerWithHealthz(t *testing.T) {
-	server := fakeHealthzServer()
-
-	go server.Run()
-	for {
-		if len(server.Address()) > 0 {
-			break
-		}
-	}
-
-	client := httptesting.New(server.Address(), false)
-
-	request := client.New(t)
-	request.Get(GogoHealthz, nil)
-	request.AssertOK()
-	request.AssertEmpty()
-}
-
-func Test_ServerWithTcp(t *testing.T) {
-	server := fakeTcpServer()
-	server.GET("/server/tcp", func(ctx *Context) {
-		ctx.SetStatus(http.StatusNotImplemented)
-	})
-
-	go server.Run()
-	for {
-		if len(server.Address()) > 0 {
-			break
-		}
-	}
-
-	client := httptesting.New(server.Address(), false)
-
-	request := client.New(t)
-	request.Get("/server/tcp", nil)
-	request.AssertStatus(http.StatusNotImplemented)
-	request.AssertEmpty()
-}
-
-var benchmarkServerWithTcpOnce sync.Once
-
-func Benchmark_ServerWithTcp(b *testing.B) {
-	server := fakeServer()
-	server.GET("/bench/tcp", func(ctx *Context) {
-		ctx.SetStatus(http.StatusNotImplemented)
-	})
-
-	var (
-		endpoint string
-	)
-	benchmarkServerWithTcpOnce.Do(func() {
-		go server.Run()
-		for {
-			if len(server.Address()) > 0 {
-				break
-			}
-		}
-
-		endpoint = "http://" + server.Address() + "/bench/tcp"
-	})
-
-	client := &http.Client{}
-
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		client.Get(endpoint)
-	}
-}
-
-func Test_ServerWithUnix(t *testing.T) {
-	it := assert.New(t)
-
-	server := fakeUnixServer()
-	server.GET("/server/unix", func(ctx *Context) {
-		ctx.SetStatus(http.StatusNotImplemented)
-	})
-
-	go server.Run()
-	for {
-		if len(server.Address()) > 0 {
-			break
-		}
-	}
-	defer os.Remove("/tmp/gogo.sock")
-
-	client := &http.Client{
-		Transport: &http.Transport{
-			DialContext: func(ctx context.Context, network, addr string) (conn net.Conn, err error) {
-				return net.Dial("unix", server.Address())
-			},
-		},
-	}
-
-	// it should work
-	response, err := client.Get("http://unix/server/unix")
-	if it.Nil(err) {
-		defer response.Body.Close()
-
-		it.Equal(http.StatusNotImplemented, response.StatusCode)
-
-		data, err := ioutil.ReadAll(response.Body)
-		it.Nil(err)
-		it.Empty(data)
-	}
-
-	// it should return error
-	response, err = http.DefaultClient.Get("http://unix/server/unix")
-	it.NotNil(err)
-	it.Nil(response)
-}
-
-var benchmarkServerWithUnix sync.Once
-
-func Benchmark_ServerWithUnix(b *testing.B) {
-	server := fakeUnixServer()
-	server.GET("/bench/unix", func(ctx *Context) {
-		ctx.SetStatus(http.StatusNotImplemented)
-	})
-
-	benchmarkServerWithUnix.Do(func() {
-		go server.Run()
-		for {
-			if len(server.Address()) > 0 {
-				break
-			}
-		}
-	})
-	defer os.Remove("/tmp/gogo.sock")
-
-	unixConn, unixErr := net.Dial("unix", "/tmp/gogo.sock")
-	client := &http.Client{
-		Transport: &http.Transport{
-			DialContext: func(ctx context.Context, network, addr string) (conn net.Conn, err error) {
-				return unixConn, unixErr
-			},
-		},
-	}
-
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		client.Get("http://unix/server/unix")
-	}
-}
 
 func Test_Server_loggerNewWithReuse(t *testing.T) {
 	it := assert.New(t)
 	logger := NewAppLogger("nil", "")
-	config, _ := fakeConfig("application.throttle.json")
+	config, _ := fakeConfig("application.throttle.yml")
 
 	server := NewAppServer(config, logger)
 
@@ -355,65 +205,6 @@ func Test_ServerWithTimeout(t *testing.T) {
 		data, err := ioutil.ReadAll(response.Body)
 		if it.Nil(err) {
 			it.Empty(data)
-		}
-	}
-}
-
-func Test_Server_Shutdown(t *testing.T) {
-	it := assert.New(t)
-
-	server := fakeTimeoutServer()
-	server.GET("/server/shutdown", func(ctx *Context) {
-		ctx.Text("SHUTDOWN")
-	})
-
-	go server.Serve()
-	for {
-		if len(server.Address()) > 0 {
-			break
-		}
-	}
-
-	// it should not work any more
-	server.Shutdown()
-
-	endpoint := "http://" + server.Address() + "/server/shutdown"
-
-	response, err := http.Get(endpoint)
-	if it.NotNil(err) {
-		it.Nil(response)
-
-		it.Contains(err.Error(), "connect: connection refused")
-	}
-}
-
-func Test_Server_ShutdownWithGhost(t *testing.T) {
-	it := assert.New(t)
-
-	server := fakeTimeoutServer()
-	server.GET("/server/shutdown/ghost", func(ctx *Context) {
-		ctx.Text("SHUTDOWN")
-	})
-
-	go server.Run()
-	for {
-		if len(server.Address()) > 0 {
-			break
-		}
-	}
-
-	// it should not work any more
-	server.Shutdown()
-
-	endpoint := "http://" + server.Address() + "/server/shutdown/ghost"
-
-	response, err := http.Get(endpoint)
-	if it.Nil(err) {
-		defer response.Body.Close()
-
-		data, err := ioutil.ReadAll(response.Body)
-		if it.Nil(err) {
-			it.Equal("SHUTDOWN", string(data))
 		}
 	}
 }
