+++ skeleton/.gitignore
@@ -27,5 +27,7 @@ _testmain.go
 *.prof
 
 # development & test config files
+*.development.yml
 *.development.json
+*.test.yml
 *.test.json
