--- skeleton/config/application.development.json
@@ -1,29 +1,0 @@
-{
-    "name": "for development",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": false,
-                "request_timeout": 3,
-                "response_timeout": 10,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            },
-            "domain": "https://example.com",
-            "getting_start": {
-                "greeting": "Hello, gogo!"
-            },
-            "debug": true
-        }
-    }
-}
