--- skeleton/config/application.healthz.json
@@ -1,23 +1,0 @@
-{
-    "name": "for healthz",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "healthz": true,
-                "ssl": false,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            }
-        }
-    }
-}
