--- skeleton/config/application.json
@@ -1,72 +1,0 @@
-{
-    "name": "gogo",
-    "mode": "test",
-    "sections": {
-        "development": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": false,
-                "request_timeout": 3,
-                "response_timeout": 10,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            },
-            "domain": "https://example.com",
-            "getting_start": {
-                "greeting": "Hello, gogo!"
-            },
-            "debug": true
-        },
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": false,
-                "request_timeout": 3,
-                "response_timeout": 10,
-                "request_id": ""
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            },
-            "domain": "https://example.com",
-            "getting_start": {
-                "greeting": "Hello, gogo!"
-            },
-            "debug": false
-        },
-        "production": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": true,
-                "ssl_cert": "/path/to/ssl/cert",
-                "ssl_key": "/path/to/ssl/key",
-                "request_timeout": 3,
-                "response_timeout": 10,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "warn",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            }
-        }
-    }
-}
