--- skeleton/config/application.tcp.json
@@ -1,22 +1,0 @@
-{
-    "name": "for tcp",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": false,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            }
-        }
-    }
-}
