--- skeleton/config/application.throttle.json
@@ -1,24 +1,0 @@
-{
-    "name": "for throttle",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "ssl": false,
-                "throttle": 1,
-                "slowdown": 1,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            }
-        }
-    }
-}
