--- skeleton/config/application.timeout.json
@@ -1,24 +1,0 @@
-{
-    "name": "for timeout",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "localhost",
-                "port": 9090,
-                "request_timeout": 1,
-                "response_timeout": 1,
-                "ssl": false,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "stderr",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            }
-        }
-    }
-}
