--- skeleton/config/application.unix.json
@@ -1,22 +1,0 @@
-{
-    "name": "for unix",
-    "mode": "test",
-    "sections": {
-        "test": {
-            "server": {
-                "addr": "unix:/tmp/gogo.sock",
-                "port": 0,
-                "ssl": false,
-                "request_id": "X-Request-Id"
-            },
-            "logger": {
-                "output": "nil",
-                "level": "debug",
-                "filter_params": [
-                    "password",
-                    "password_confirmation"
-                ]
-            }
-        }
-    }
-}
