+++ skeleton/config/application.yml
@@ -1,0 +1,47 @@
+---
+mode: test
+name: gogo
+
+default_server: &default_server
+  addr: localhost
+  port: 9090
+  ssl: false
+  request_timeout: 3
+  response_timeout: 10
+  request_id: X-Request-Id
+
+default_logger: &default_logger
+  output: nil
+  level: debug
+  filter_params:
+    - password
+    - password_confirmation
+
+sections:
+  development:
+    server:
+      <<: *default_server
+    logger:
+      <<: *default_logger
+    domain: https://example.com
+    getting_start:
+      greeting: Hello, gogo!
+    debug: true
+  test:
+    server:
+      <<: *default_server
+      request_id: ''
+    logger:
+      <<: *default_logger
+    domain: https://example.com
+    getting_start:
+      greeting: Hello, gogo!
+    debug: false
+  production:
+    server:
+      <<: *default_server
+      ssl: true
+      ssl_cert: "/path/to/ssl/cert"
+      ssl_key: "/path/to/ssl/key"
+    logger:
+      <<: *default_logger
